image, refetch, select, css 수정

+++ src/components/units/Admin/Admin.queries.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const FETCH_CAR_REGISTRATIONS = gql`
-  query fetchCarRegistrations($page: Int!) {
+  query fetchCarRegistrations($page: Float!) {
     fetchCarRegistrations(page: $page) {
       id
       carNumber
